Finished class Config and BusinessEvent. To do sth with diff event.

+++ atomservice.h
@@ -13,7 +13,7 @@ public:
     bool free;
 
     int price;
-    double reliability;
+    int reliability;
     int execTime;
 
     bool operator ==(AtomService &other);
